pyproject.toml
[tool.poetry]
name = "solicitudes-cai"
version = "0.1.0"
description = "Plataforma de solicitudes del Centro de Alumnos de Ingeniería"
authors = ["Centro de Alumnos de Ingeniería <contacto@cai.cl>"]
maintainers = ["Agustín Covarrubias <agucova@uc.cl>"]
readme = "README.md"
packages = [
    { include = "cai" },
    { include = "accounts" },
    { include = "activities" },
]

[tool.poetry.dependencies]
python = "3.10.12"
django = { extras = ["argon2"], version = "^4.2.3" }
python-dotenv = "^1.0.0"
requests = "^2.31.0"
sendgrid = "^6.10.0"
psycopg2-binary = "^2.9.6"
dj-database-url = "^2.0.0"
gunicorn = "^20.1.0"
whitenoise = { extras = ["brotli"], version = "^6.5.0" }


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.277"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
# See https://beta.ruff.rs/docs/rules/ for more information
select = [
    "E",
    "F",
    "NPY",
    "PD",
    "N",
    "C90",
    "C4",
    "PT",
    "SIM",
    "TCH",
    "UP",
    "ICN",
    "PIE",
    "ERA",
    "TRY",
    "RUF",
    "B",
    "TD",
    "DJ",
]

# Disabled rules
# TRY003 (long messages in exceptions) is just annoying
# TD002 requires authors in TODOs, which seems unnecessary
# TD003 requires links, but somethimes this is too bulky
ignore = ["TRY003", "TD002", "TD003"]

# Migration files are autogenerated, so we don't need to check them
extend-exclude = ["migrations"]

[tool.ruff.mccabe]
max-complexity = 10

[tool.black]
line-length = 120
target-version = ["py310"]